完成老师部分

+++ score-common/src/main/java/com/middleschool/score/common/dto/MsScore.java
@@ -7,7 +7,7 @@ public class MsScore {
 
     private Long studentId;
 
-    private Integer classId;
+    private Long classId;
 
     private Integer semester;
 
@@ -49,11 +49,11 @@ public class MsScore {
         this.studentId = studentId;
     }
 
-    public Integer getClassId() {
+    public Long getClassId() {
         return classId;
     }
 
-    public void setClassId(Integer classId) {
+    public void setClassId(Long classId) {
         this.classId = classId;
     }
